Fixed error, spring boot not starting. Renamed user to person and changs in application properties. Commented fields that involves relationships. This will be fixed in other branch
