+++ src/main/java/se/iths/grocerylistgenerator/model/Ingredient.java
@@ -12,13 +12,19 @@ public class Ingredient {
     private String name;
     private String quantity;
 
-    @ManyToOne
-    private Category category;
+//    @ManyToOne
+//    private Category category;
 
-    public Ingredient(String name, String quantity, Category category) {
+//    public Ingredient(String name, String quantity, Category category) {
+//        this.name = name;
+//        this.quantity = quantity;
+////        this.category = category;
+//    }
+
+
+    public Ingredient(String name, String quantity) {
         this.name = name;
         this.quantity = quantity;
-        this.category = category;
     }
 
     public Ingredient() {
@@ -48,11 +54,11 @@ public class Ingredient {
         this.quantity = quantity;
     }
 
-    public Category getCategory() {
-        return category;
-    }
-
-    public void setCategory(Category category) {
-        this.category = category;
-    }
+//    public Category getCategory() {
+//        return category;
+//    }
+//
+//    public void setCategory(Category category) {
+//        this.category = category;
+//    }
 }
