+++ src/main/java/se/iths/grocerylistgenerator/model/Store.java
@@ -11,13 +11,18 @@ public class Store {
 
     private String name;
 
-    @OneToOne(cascade = CascadeType.ALL)
-    @JoinColumn(name = "storecontactinfo_id")
-    private StoreContactInfo storeContactInfo;
+//    @OneToOne(cascade = CascadeType.ALL)
+//    @JoinColumn(name = "storeinfo_id")
+//    private StoreContactInfo storeContactInfo;
 
-    public Store(String name, StoreContactInfo storeContactInfo) {
+//    public Store(String name, StoreContactInfo storeContactInfo) {
+//        this.name = name;
+//        this.storeContactInfo = storeContactInfo;
+//    }
+
+
+    public Store(String name) {
         this.name = name;
-        this.storeContactInfo = storeContactInfo;
     }
 
     public Store() {
@@ -39,11 +44,11 @@ public class Store {
         this.name = name;
     }
 
-    public StoreContactInfo getStoreContactInfo() {
-        return storeContactInfo;
-    }
-
-    public void setStoreContactInfo(StoreContactInfo storeContactInfo) {
-        this.storeContactInfo = storeContactInfo;
-    }
+//    public StoreContactInfo getStoreContactInfo() {
+//        return storeContactInfo;
+//    }
+//
+//    public void setStoreContactInfo(StoreContactInfo storeContactInfo) {
+//        this.storeContactInfo = storeContactInfo;
+//    }
 }
