+++ src/main/resources/application.properties
@@ -4,4 +4,5 @@ spring.datasource.username=postgres
 spring.datasource.password=postgres
 spring.jpa.show-sql=true
 spring.jpa.properties.hibernate.dialect = org.hibernate.dialect.PostgreSQLDialect
-spring.jpa.hibernate.ddl-auto = create
+spring.jpa.hibernate.ddl-auto = create-drop
+spring.jpa.properties.hibernate.format_sql=true
